requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python-headless
tensorflow-cpu
